Major OSLua Update
- Successfully added Multitasking (Start a program and leave with
Delete, if you open the app again it will be at the same state as the
last time you opened it (next goal is a taskmanager, which let's you
kill your processes))

- Optimized AppSearch
INFO:
Apps which need to be at a specific directory (e.g. Quest (by oeed)),
won't launch properly. A temporaly fix is to move the actual program to
where it needs to be and write in a startup file in the .app folder,
which launches the actual program:
quest --> /quest
startup --> /doorOS/apps/Quest.app/startup

startup:
shell.run("/quest")

--> every function still works (e.g. exiting the app)